--- src/.dockerignore
@@ -1,13 +1,0 @@
-*
-!protos
-!cmd
-!pkg
-!go.mod
-!go.sum
-cmd/server/main
-cmd/server/server
-cmd/cli
-!cd
-cd/node_modules
-tests
-**/*_test.go
